backend/flask-api/requirements.txt
gunicorn==19.7
python-decouple==3.3
whitenoise==3.3.1
psycopg2==2.8.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-RESTful==0.3.8
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-HTTPAuth==4.1.0
beautifulsoup4==4.9.3
lxml==4.6.2
Flask-Mail==0.9.1
passlib==1.7.4
itsdangerous==1.1.0
WeasyPrint==52.2
python-decouple==3.3